kleaf: Fix ddk_headers_archive when @kleaf is a submodule.
If @kleaf is a submodule, previously, @kleaf//common:all_headers_aarch64
fails to build. Fix it by fixing ddk_headers_archive to handle paths
correctly.

Also, the generated archive will now consistently drop the workspace
root. This ensures that, no matter if @kleaf is a submodule or not, the
generated archive contains common/ as a top-level directory.

Bug: 377799264
Change-Id: Ic996d1b5bb5d23b9f73c410ddb71bab52ba349a3